Rotation and angular momentum now in object list
As indices 5 and 6 respectively.
Timeleft is now index 7.
Also fixes a pgx bug that erased comments in writing to the line, now filehelper checks for a comment (must start with #) and doesn't write over it.
However, if you wish to write over everything there is an optional argument called override that will do that