saves with pickle
I ran into weird issues saving each sector to an individual line, so it saves all the sectors in one large list as a single object at the end of gamedata.
For the curious, it uses pickle protocol-0 to create a string which is then encoded in base64 because compression, not that it really needs it.
Right now it's not very efficient with save and load, but it seems like it works and I'll fix it later.